Refactor content links
